requirements.txt
ezdxf==1.0.3
matplotlib==3.7.2
pandas==2.0.3
pyodbc==4.0.39
python-dotenv==1.0.0
sqlalchemy==2.0.20